chore (1/2): update the production workflow
Call pandoc once only, to create the PDF output. Intermediate Markdown
files will no longer be written to the `build` directory. The contents
of `build` are henceforth ignored.

Markdown-formatted schedules must now be committed to the repository,
until I can reproduce the CSV-to-Markdown conversion with GitHub
Actions. `.gitignore` is updated to watch these files.

The GitHub Action is updated to supply a new requirement of the remote
build. (Previously this Lua filter was employed locally only.)

The README is updated to reflect this new production workflow etc.